feat(o-shop): separated ProductForm and ProductCard as components, redirect to admin-product on save
