+++ src/styles/styles.scss
@@ -14,3 +14,7 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
+
+#productcard {
+  width: 18rem;
+}
